Kabo Latest Updates 31 July 2024
User entity update to take image for license as well,

User Service Tests currently working

The front end currently registers a user successfully to the database

need to debugg the registration form to see why images are showing as null in the database.

Simply need to add a password variable that will be user to test when logging in. this i can add to the contact class paired with email